Added ExploreModePlanet script
Added and changed scripts to help integrate data persistence into ExploreMode2D scene
