+++ LoneMiner/Assets/Scripts/ExploreModePlanet.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2ea287a765000994f9ba7403d43bc539
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
