+++ LoneMiner/Assets/Scripts/ExploreModePlayerControls.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 16fb1e3c37d2e0b48b22ffd58dfe4f7b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
