+++ LoneMiner/ProjectSettings/ProjectVersion.txt
@@ -1,1 +1,1 @@
-m_EditorVersion: 2018.2.15f1
+m_EditorVersion: 2018.2.16f1
